requirements.txt
hydra-score
higher
wandb
datasets
peft
omegaconf
jsonlines